json-exports: renamed exports, made duplicate postgres functions static
