+++ src/exports.c
@@ -1,4 +1,4 @@
-#include <json-exports.h>
+#include <exports.h>
 
 typedef struct Needles *Needles;
 typedef struct Needles {
